Reduce query load by using .values()
The tricky part here was in dealing with the related videos - we still
have to do a bit of testing to see if there's no way to keep the
prefetch as it existed rather than doing the manual join in our app
code.

We've also lost some slight distinction in Consent Ruling comments
between "no previous ruling" and "no comment in previous ruling" due to
the nature of the optimized query.