Always use PNG favicon with redirect view
In the old clusters, we had an nginx reverse proxy handling the path for
the favicon. To replicate this old functionality the same way, we'd need
to create an entirely new ingress rule to redirect to the old favicon.ico file.

Considering the fact that we already have the PNG in our staging bucket,
we're just going to let django handle this redirect.